Replace Visual Studio solution with CMake
The goal of this change is to enhance portability by switching to a
cross-platform build system.

The current configuration produces output that is almost identical to
the original project files. However, the Doom3BFG and Game-d3xp projects
have been combined. This was necessary because the executable project
used a linker property that caused it to link against the object files
of its dependent projects as opposed to the resulting library files.
This doesn't seem to make a difference with the idLib or doomclassic
libraries yet Game-d3xp caused numerous issues when it was linked in a
traditional manner. The easiest (only?) way to emulate the "Use Library
Dependency Inputs" behaviour in CMake is to combine the projects.